Ensure storage pool tests use an empty directory
All tests involving a storage pool used `/mnt` has the hard-coded source
for the pool. This fails if the hosts `/mnt` is *not* empty. This might
actually not be true and in some cases (e.g. Github Actions VM) it
cannot be changed at all by the developer.

This CL changes all tests to use a new temporary directory, which
*always* is empty as the pool source.